test for LC sim now in separate file
